streamlit_app/requirements.txt
streamlit 
pandas
numpy
pillow
tensorflow-cpu